refactor(record): organize attachment migrations by family
