+++ packages/niceeval/src/record/family/index.ts
@@ -1,6 +1,6 @@
 export * from "./common.ts";
-export * from "./assertions.ts";
-export * from "./observability.ts";
+export * from "./assertions/definition.ts";
+export * from "./observability/definition.ts";
 export * from "./file-changes.ts";
 export * from "./sources.ts";
 export * from "./artifacts.ts";
